implemented gndlink.js
For a given PND number this obtains an  JSON array of links (URL + label)
as provided by Histvv::Apache::GNDLink and adds those as hyperlinks to the
"Links" section of a page. Links are not added if there already is a link
with the same link text. The script relies on the variables 'gndlinkURL'
and 'gndlinkPndPara' to be set.